adk/requirements.txt
fastapi
uvicorn
requests
firebase-admin
google-generativeai
python-dotenv
